refactor: move table widgets to different files
This is just some simple refactoring to make things a bit cleaner.
