+++ src/app/widgets.rs
@@ -1,2 +1,8 @@
-pub mod process_widget;
-pub use process_widget::*;
+pub mod process_table_widget;
+pub use process_table_widget::*;
+
+pub mod temperature_table_widget;
+pub use temperature_table_widget::*;
+
+pub mod disk_table_widget;
+pub use disk_table_widget::*;
